requirements.txt
colorama==0.4.6
dnspython==2.7.0
requests==2.32.3
wireguard-tools==0.5.0
